refactor: group components
