+++ src/features/MainScreen/NotesOverview/index.tsx
@@ -13,7 +13,7 @@ import { Icon } from '@components/Icon/Icon.bundle/common';
 import { List } from '@components/List';
 import { TagEditor, TagEditorData } from '@components/TagEditor';
 import { IResolvedTag } from '@core/features/tags';
-import { useTagsRegistry } from '@features/Workspace/WorkspaceProvider';
+import { useTagsRegistry } from '@features/App/Workspace/WorkspaceProvider';
 import { useAppDispatch } from '@state/redux/hooks';
 import { useWorkspaceData, useWorkspaceSelector } from '@state/redux/profiles/hooks';
 import {
