+++ src/features/MonakoEditor/features/useEditorLinks.ts
@@ -3,7 +3,7 @@ import { saveAs } from 'file-saver';
 import { editor, languages } from 'monaco-editor-core';
 import { findLinksInText, getAppResourceDataInUrl } from '@core/features/links';
 import { openLink } from '@electron/requests/interactions/renderer';
-import { useFilesRegistry } from '@features/Workspace/WorkspaceProvider';
+import { useFilesRegistry } from '@features/App/Workspace/WorkspaceProvider';
 
 import { useNotesControl } from '../../MainScreen/useNotesControl';
