+++ src/features/NoteEditor/index.tsx
@@ -34,7 +34,7 @@ import {
 	useAttachmentsController,
 	useFilesRegistry,
 	useTagsRegistry,
-} from '@features/Workspace/WorkspaceProvider';
+} from '@features/App/Workspace/WorkspaceProvider';
 import { useAppDispatch } from '@state/redux/hooks';
 import { useWorkspaceData, useWorkspaceSelector } from '@state/redux/profiles/hooks';
 import { selectTags, workspacesApi } from '@state/redux/profiles/profiles';
