+++ src/hooks/notes/useCreateNote.ts
@@ -1,6 +1,9 @@
 import { useCallback, useEffect, useRef } from 'react';
 import { NoteId } from '@core/features/notes';
-import { useNotesRegistry, useTagsRegistry } from '@features/Workspace/WorkspaceProvider';
+import {
+	useNotesRegistry,
+	useTagsRegistry,
+} from '@features/App/Workspace/WorkspaceProvider';
 import { useWorkspaceSelector } from '@state/redux/profiles/hooks';
 import { selectActiveTag, selectNotes } from '@state/redux/profiles/profiles';
